代码优化，结构调整，进度50%
