+++ src/layout/components/Navbar.vue
