+++ src/layout/mixin/ResizeHandler.js
