+++ src/views/dashboard/BarChart.vue
