+++ src/views/features/401.vue
@@ -4,7 +4,7 @@
       <el-col :span="12">
         <h1 class="text-jumbo text-ginormous">Oops!</h1>
         图片来源<a href="https://zh.airbnb.com/" target="_blank">airbnb</a> 页面
-        <h2>你没有权限去该页面</h2>
+        <h2>你没有权限去操作</h2>
         <h6>如有不满请联系你领导</h6>
         <ul class="list-unstyled">
           <li>或者你可以去:</li>
