+++ src/views/tools/email/config.vue
